requirements.txt
pymongo
matplotlib
pandas
motor
coloredlogs
youtube_dl
praw
discord.py
PyNaCl
dnspython